Break async operation runner tests into separate files
